+++ src/Middleware/BaseMiddleware.php
@@ -11,14 +11,18 @@
 namespace SebLucas\Cops\Middleware;
 
 use SebLucas\Cops\Handlers\BaseHandler;
+use SebLucas\Cops\Input\HasContextInterface;
+use SebLucas\Cops\Input\HasContextTrait;
 use SebLucas\Cops\Input\Request;
 use SebLucas\Cops\Output\Response;
 
 /**
  * Summary of BaseMiddleware
  */
-abstract class BaseMiddleware
+abstract class BaseMiddleware implements HasContextInterface
 {
+    use HasContextTrait;
+
     /**
      * @param Request $request
      * @param BaseHandler $handler